Add RequirementsManagerSubApp
