+++ __init__.py
@@ -4,18 +4,17 @@ from tkinter import ttk, simpledialog, filedialog, scrolledtext
 
 from .mainappsrc.AutoML import (
     AutoMLApp,
-    PMHF_TARGETS,
     FaultTreeNode,
     AutoML_Helper,
     messagebox,
     GATE_NODE_TYPES,
 )
+from config.automl_constants import PMHF_TARGETS
 from analysis.models import HazopDoc
 from gui.dialogs.edit_node_dialog import EditNodeDialog
 
 __all__ = [
     "AutoMLApp",
-    "PMHF_TARGETS",
     "FaultTreeNode",
     "AutoML_Helper",
     "messagebox",
@@ -25,6 +24,7 @@ __all__ = [
     "filedialog",
     "scrolledtext",
     "GATE_NODE_TYPES",
+    "PMHF_TARGETS",
     "HazopDoc",
     "EditNodeDialog",
 ]
